+++ object.h
@@ -28,7 +28,7 @@ public:
 	*		Hit(ray, hitrec, 0.001, DBL_MAX);
 	*/
 	virtual bool Hit(const Ray &r, HitRecord &rec, double tMin, double tMax) const = 0;
-	virtual int clType() const = 0;
+	virtual double clType() const = 0;
 	virtual Vector3D clCenter() const = 0;
 	virtual double clRadius() const = 0;
 	virtual Vector3D clBound1() const = 0;
